apps/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.18.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.0
python-multipart==0.0.6
pillow==10.1.0
